Add files via upload
split the testcase, include a new test document
