+++ Sprint3_Zhonghua_Bao/Project01_Zhonghua_Bao.ged
@@ -1,0 +1,157 @@
+0 NOTE Zhonghua Bao
+0 HEAD
+1 SOUR Family Echo
+2 WWW http://www.familyecho.com/
+1 FILE My Family
+1 DATE 28 JAN 2019
+1 DEST ANSTFILE
+1 GEDC
+2 VERS 5.5.1
+2 FORM LINEAGE-LINKED
+1 SUBM @I1@
+2 NAME Zhonghua Bao
+1 SUBN
+1 CHAR UTF-8
+0 @I1@ INDI
+1 NAME Zhonghua /Tang/
+2 GIVN Zhonghua
+2 SURN Tang
+2 _MARNM Bao
+1 SEX M
+1 BIRT
+2 DATE 5 OCT 1992
+1 FAMC @F1@
+0 @I2@ INDI
+1 NAME Songhua /Yu/
+2 GIVN Songhua
+2 SURN Yu
+2 _MARNM Yu
+1 SEX M
+1 BIRT
+2 DATE 26 MAR 1967
+1 FAMS @F1@
+1 FAMC @F2@
+0 @I3@ INDI
+1 NAME Jun /Bao/
+2 GIVN Jun
+2 SURN Bao
+2 _MARNM Bao
+1 SEX F
+1 BIRT
+2 DATE 29 NOV 1968
+1 FAMS @F1@
+1 FAMS @F3@
+1 FAMC @F4@
+0 @I4@ INDI
+1 NAME Yuxing /Bao/
+2 GIVN Yuxing
+2 SURN Bao
+2 _MARNM Bao
+1 SEX F
+1 BIRT
+2 DATE 7 DEC 1940
+1 FAMS @F4@
+0 @I5@ INDI
+1 NAME Hongzi /Zhou/
+2 GIVN Hongzi
+2 SURN Zhou
+2 _MARNM Zhou
+1 SEX M
+1 BIRT
+2 DATE 23 FEB 1941
+1 FAMS @F4@
+0 @I6@ INDI
+1 NAME Guanghui /Tang/
+2 GIVN Guanghui
+2 SURN Tang
+2 _MARNM Tang
+1 SEX M
+1 BIRT
+2 DATE 6 JUL 1964
+1 FAMS @F3@
+0 @I7@ INDI
+1 NAME Qiwen /Tang/
+2 GIVN Qiwen
+2 SURN Tang
+2 _MARNM Tang
+1 SEX M
+1 BIRT
+2 DATE 12 MAY 1991
+1 FAMC @F3@
+0 @I8@ INDI
+1 NAME Cungen /Yu/
+2 GIVN Cungen
+2 SURN Yu
+2 _MARNM Yu
+1 SEX M
+1 BIRT
+2 DATE 23 DEC 1939
+1 FAMS @F2@
+0 @I9@ INDI
+1 NAME Liying /Cai/
+2 GIVN Liying
+2 SURN Cai
+2 _MARNM Cai
+1 SEX F
+1 BIRT
+2 DATE 7 APR 1944
+1 FAMS @F2@
+0 @I10@ INDI
+1 NAME Qin /Bao/
+2 GIVN Qin
+2 SURN Bao
+2 _MARNM Bao
+1 SEX F
+1 BIRT
+2 DATE 15 NOV 1973
+1 FAMC @F4@
+0 @I11@ INDI
+1 NAME Jianping /Yu/
+2 GIVN Jianping
+2 SURN Yu
+2 _MARNM Yu
+1 SEX F
+1 BIRT
+2 DATE 5 FEB 1962
+1 FAMC @F2@
+0 @I12@ INDI
+1 NAME Songnian /Yu/
+2 GIVN Songnian
+2 SURN Yu
+2 _MARNM Yu
+1 SEX M
+1 BIRT
+2 DATE 5 FEB 1962
+1 FAMC @F2@
+0 @F1@ FAM
+1 HUSB @I2@
+1 WIFE @I3@
+1 CHIL @I1@
+1 MARR
+2 DATE 15 SEP 1997
+2 PLAC Shanghai
+1 _CURRENT Y
+0 @F3@ FAM
+1 HUSB @I6@
+1 WIFE @I3@
+1 CHIL @I7@
+1 MARR
+2 DATE 20 APR 1990
+2 PLAC Shanghai
+1 DIV
+2 DATE 11 SEP 1995
+1 _CURRENT N
+0 @F2@ FAM
+1 HUSB @I8@
+1 WIFE @I9@
+1 CHIL @I2@
+1 CHIL @I11@
+1 CHIL @I12@
+1 _CURRENT Y
+0 @F4@ FAM
+1 HUSB @I4@
+1 WIFE @I5@
+1 CHIL @I3@
+1 CHIL @I10@
+1 _CURRENT Y
+0 TRLR
