+++ Sprint3_Zhonghua_Bao/UserStory21.py
@@ -1,6 +1,3 @@
-import unittest
-import Project03_SourceCode
-
 def userStory21(result):
     indi,fam = result['indi'],result['fam']
     anamoly = list()
@@ -24,14 +21,3 @@ def userStory21(result):
                 print("ERROR: INDIVIDUAL: US21:", check_wife, "should be a female.")
     
     return anamoly
-
-class UserStory21Test(unittest.TestCase):
-    def test_userStory21(self):
-        r = Project03_SourceCode.parse_file('Project03-TestFile.ged')
-        expect = ['I2']
-        self.assertEqual(expect,userStory21(r))
-
-
-
-if __name__ == '__main__':
-    unittest.main(verbosity=2)
