+++ Sprint3_Zhonghua_Bao/UserStory32.py
@@ -1,6 +1,3 @@
-import unittest
-import Project03_SourceCode
-
 def userStory32(result):
     indi,fam = result['indi'],result['fam']
     anamoly = list()
@@ -18,14 +15,3 @@ def userStory32(result):
                     tmp.add(birth)
     
     return anamoly
-
-class UserStory32Test(unittest.TestCase):
-    def test_userStory32(self):
-        r = Project03_SourceCode.parse_file('Project03-TestFile.ged')
-        expect = ['F2']
-        self.assertEqual(expect,userStory32(r))
-
-
-
-if __name__ == '__main__':
-    unittest.main(verbosity=2)
